requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-include-by-ajax==2.0.0
django-js-asset==1.2.2
django-taggit==1.3.0
django-user-agents==0.4.0
docutils==0.15.2
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
ua-parser==0.10.0
user-agents==2.1python-dotenv
